+++ src/libawkward/builder/DatetimeBuilder.cpp
@@ -42,7 +42,7 @@ namespace awkward {
     content_.concatenate(
       reinterpret_cast<int64_t*>(
         container.empty_buffer(form_key.str() + "-data",
-        content_.length() * (int64_t)sizeof(int64_t))));
+        (int64_t)content_.length() * (int64_t)sizeof(int64_t))));
 
     std::string primitive(units_);
